requirements.txt
# Core Pathway Framework (MANDATORY)
pathway>=0.8.0

# LLM and Embeddings
openai>=1.0.0
anthropic>=0.18.0
groq>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=2.0.0

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.8.0

# Vector Store and Retrieval
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Natural Language Inference
datasets>=2.16.0

# Graph Processing for Causal Chains
networkx>=3.2.0

# Temporal Processing
dateparser>=1.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0

# Optional: For advanced reasoning
langchain>=0.1.0
langchain-community>=0.0.10

# Logging
loguru>=0.7.0
